# Optimize print table