This pull request updates the GPU benchmarking suite to improve usability, reporting, and test configuration. The main changes include reducing default test durations and intervals, increasing matrix sizes and precision for compute tests, refining memory bandwidth tests, and replacing verbose printouts with a live summary table.

**Benchmark configuration and test improvements:**

* Reduced the default test duration from 10 to 5 seconds and the test interval from 30 to 10 seconds across the CLI, `GPUBenchmark` class, and `run_gpu_benchmark` function for faster and more frequent benchmarking cycles. [[1]](diffhunk://#diff-9e3faf088c7a642e1fa397625c7edfadf7b2f411cbb70a1e6f5d7933006e32faL217-R225) [[2]](diffhunk://#diff-8da985c52d6103a0ff24a2970c06ba68582e43157c8237ed2b207a2c144f5ff7L22-R22) [[3]](diffhunk://#diff-8da985c52d6103a0ff24a2970c06ba68582e43157c8237ed2b207a2c144f5ff7L273-R299)
* Increased matrix size in the compute benchmark from 4096 to 16,384 and switched to `torch.bfloat16` for higher performance and more realistic stress testing. Also added reporting of the matrix data type. [[1]](diffhunk://#diff-3d028aa42341286daf640e5d03bed313134f537f87cf6deca05bc165f0e622c2L35-R38) [[2]](diffhunk://#diff-3d028aa42341286daf640e5d03bed313134f537f87cf6deca05bc165f0e622c2R72)
* Refined the memory bandwidth test to use half the previous data size (128M elements), always operate on CPU-to-GPU transfers, and dynamically determine element size. Also added reporting of memory used and data type. [[1]](diffhunk://#diff-92849ced70abee2e7ce4a014cb146ea3854502a01300016258de03eb418402f2L30-R43) [[2]](diffhunk://#diff-92849ced70abee2e7ce4a014cb146ea3854502a01300016258de03eb418402f2R54-R63)

**Reporting and output enhancements:**

* Replaced verbose per-test printouts with a live, formatted summary table in the main thread showing key metrics for each GPU, including TFLOPs, matrix size, memory usage, bandwidth, and last test time. The table updates every second and uses ANSI escape codes to clear the screen for a dashboard-like experience. [[1]](diffhunk://#diff-8da985c52d6103a0ff24a2970c06ba68582e43157c8237ed2b207a2c144f5ff7L176-R247) [[2]](diffhunk://#diff-8da985c52d6103a0ff24a2970c06ba68582e43157c8237ed2b207a2c144f5ff7L160-L162) [[3]](diffhunk://#diff-8da985c52d6103a0ff24a2970c06ba68582e43157c8237ed2b207a2c144f5ff7L132-L142) [[4]](diffhunk://#diff-8da985c52d6103a0ff24a2970c06ba68582e43157c8237ed2b207a2c144f5ff7L250-R276)

**Other improvements:**

* Adjusted subprocess call for memory tests to use the configurable test duration and timeout, ensuring consistency with user settings.